# backend/docs/swagger.yaml
openapi: 3.0.3
info:
  title: GameHub API
  description: GameHub游戏社区平台API文档
  version: 1.0.0
  contact:
    name: GameHub开发团队
    email: dev@gamehub.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: 本地开发服务器
  - url: https://api.gamehub.example.com/api/v1
    description: 生产服务器
...
paths:
...
  /follow/{userId}:
    post:
      tags:
        - 关注
      summary: 关注用户
      description: 关注其他用户
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: 关注成功
    delete:
      tags:
        - 关注
      summary: 取消关注
      description: 取消对其他用户的关注
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 取消关注成功
...
  /follow/{userId}/followers:
    get:
      tags:
        - 关注
      summary: 获取粉丝列表
      description: 获取用户的粉丝列表
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: 获取成功
  /follow/{userId}/following:
    get:
      tags:
        - 关注
      summary: 获取关注列表
      description: 获取用户关注的用户列表
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: 获取成功
...
  /follow/stats:
    get:
      tags:
        - 关注
      summary: 关注统计
      description: 获取当前用户的关注统计
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      followingCount:
                        type: integer
                      followersCount:
                        type: integer
  /follow/mutual/{userId}:
    get:
      tags:
        - 关注
      summary: 检查是否互相关注
      description: 检查当前用户与指定用户是否互相关注
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      isMutual:
                        type: boolean
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...